Added forced download of .txt, .do, and .log files
This is a first pass and seems to be working. The function is based heavily on the default download. The result is that it works, but there are six added import statements.

Is there a better way?

Atleast one, from ckan.common import request, could be combined with an existing import statement - from ckan.common import ...

How else could the organization be improved?