Update pop-up info boxes in Daraforms.
This also includes adding a pop-up window for the 'File' field in the data form. Previously, this field had an icon for info, but clicking on it didn't do anything. To get that working, the 'image_upload' macrot was copied from ckan/templates/macros/form.html to dara/templates/macros/dara_form.html so it could be modified to take 'attrs' as a parameter to pass to the form.input macro.