Revise author in resource form
Lookups working for last name and institution. Turned off for first name.
IDs for authors and institutions -- TODO: make sure that IDs are blank if a name is typed rather than selected from an option.
If a name and instition are given in the same section, the institution is treated as the affiliation in the XML.