Fix error with dates giving error if there was an error somewhere else in the form.
Any error in the form was causing the dates to throw an error too. Even if the fields were blank, a message would appear sayign that they are required. This was caused by how the error handling was set up in the emplate (dara_form.html). Now all error handling for the 'dates' field is handled by the javascript function