Fix resources being created when press "Create dataset"
When the user arrives at the step to create a resource for a dataset,
they have the option to go back to the dataset form by clicking a button.
Due to how the dara forms are set up, with some fields being auto popuated
and required, pressing the buttong would trigger the creation of the resourc
rather than just going back.

Normally, it would go back a step as expected if there was no data, and if there
was data added the resource would be created. That is being bypassed in order
to go back a step. Nothing could be done with respect to the form being emptied.
so it's not possible to get it working as intended without an overhaul that might
make the form less user friendly - some of the fields are pre-populated with
values from the previous dataset form so the user doesn't need to repeat information
unnecessarily.

-- Also fixes the "cancel" function to work with the new url structure